+++ algorithms/shared_ppo.py
@@ -6,7 +6,7 @@ from all.logging import DummyWriter
 from all.presets import IndependentMultiagentPreset, Preset
 from all.core import State
 import torch
-from env_utils import make_env
+from env_utils import make_env, make_vec_env
 import supersuit as ss
 from models import impala_features, impala_value_head, impala_policy_head, nature_features
 from env_utils import InvertColorAgentIndicator
@@ -14,27 +14,8 @@ from all.bodies import DeepmindAtariBody
 from models import ImpalaCNNLarge
 from all import nn
 
-
-def make_vec_env(env_name, device):
-    import importlib
-    env = importlib.import_module('pettingzoo.atari.{}'.format(env_name)).parallel_env(obs_type='grayscale_image')
-    env = ss.max_observation_v0(env, 2) # stacking observation: (env, 2)==stacking 2 frames as observation
-    env = ss.frame_skip_v0(env, 4) # frame skipping: (env, 4)==skipping 4 or 5 (randomly) frames
-    # env = InvertColorAgentIndicator(env) # handled by body
-    env = ss.resize_v0(env, 84, 84) # resizing
-    env = ss.reshape_v0(env, (1, 84, 84)) # reshaping (expand dummy channel dimension)
-    env = ss.black_death_v2(env) # Give black observation (zero array) and zero reward to dead agents
-    env = InvertColorAgentIndicator(env)
-    # env = to_parallel(env)
-    env = ss.pettingzoo_env_to_vec_env_v0(env)
-    env = ss.concat_vec_envs_v0(env, 32, num_cpus=8, base_class='stable_baselines3')
-    env = GymVectorEnvironment(env, env_name, device=device)
-    return env
-
-
 def nat_features():
     return nature_features(16)
-
 
 def make_ppo_vec(env_name, device, _):
     venv = make_vec_env(env_name, device)
